- More precision in FPS Control
git-svn-id: https://svn.code.sf.net/p/bennugd/code@100 3fa62efa-c05c-4319-b09d-9b630d73c7a0

+++ modules/librender/g_frame.c
@@ -139,11 +139,11 @@ void gr_wait_frame()
 
         fps_partial = 1000.0 / ticks_per_frame ;
 
-        if ( ( ( int ) fps_partial ) == fps_value )
+        if ( ( /*( int )*/ fps_partial ) == fps_value )
         {
             jump = 0;
         }
-        else if ( ( ( int ) fps_partial ) > fps_value )
+        else if ( ( /*( int )*/ fps_partial ) > fps_value )
         {
             int delay = fps_partial * frame_ms - fps_value * frame_ms ;
